ambient-email-agent/requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langgraph>=0.1.0
